==> 04_picasso_add_feature.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns
sns.set()
import re

# load data
df = pd.read_csv('picasso_data_before_text.csv')
cols = ['item_description', 'item_title', 'auc_end_weekday',
       'sold_price_value_usd', 'est_low_value_usd', 'est_high_value_usd',
       'bid_start_value_usd', 'auc_to_now_days', 'auc_region',
       'item_title_word_count', 'item_description_word_count']
df.columns = cols

# re-arrange columns
df = df[[
    'sold_price_value_usd',
    'bid_start_value_usd', 'est_low_value_usd', 'est_high_value_usd',
    'auc_to_now_days',
    'auc_region',
    'auc_end_weekday',
    'item_title_word_count', 'item_description_word_count',
    'item_title', 'item_description'   
]]

# lower text
df['item_title'] = df['item_title'].apply(lambda x: x.lower())
df['item_description'] = df['item_description'].apply(lambda x: x.lower())

# get signature info

def get_sign(string):
    if string.find('sign') == -1:
        return 0
    else:
        return 1

df['sign'] = ((df['item_description'].apply(get_sign) + df['item_title'].apply(get_sign) + 1) / 2).apply(int)

def signed_or_not(number):
    if number == 1:
        return 'signed'
    else:
        return 'not_signed'

df['sign'] = df['sign'].apply(signed_or_not)

# get lithograph info

def get_litho(string):
    if string.find('litho') == -1:
        return 0
    else:
        return 1

df['litho'] = ((df['item_description'].apply(get_litho) + df['item_title'].apply(get_litho) + 1) / 2).apply(int)

def get_litho(string):
    if string.find('litho') == -1:
        return 0
    else:
        return 1

df['litho'] = ((df['item_description'].apply(get_litho) + df['item_title'].apply(get_litho) + 1) / 2).apply(int)

def litho_or_not(number):
    if number == 1:
        return 'litho'
    else:
        return 'not_litho'

df['litho'] = df['litho'].apply(litho_or_not)

df3.to_csv('picasso_data_for_model.csv', mode='a', index=False, header=False)

# headers = ['sold_price_value_usd', 'bid_start_value_usd', 'est_low_value_usd',
#        'est_high_value_usd', 'auc_to_now_days', 'auc_region',
#        'auc_end_weekday', 'item_title_word_count',
#        'item_description_word_count', 'item_title', 'item_description', 'sign',
#        'litho']
